refactor: move the IUnitOfWork interface to the business layer
